compose fiber tree primitives
